Refactor DistrictMap component
